restructure schema package
